축제관리 이미지 추가 기능 개선

+++ Front/project_Full/src/js/pages/Festivals/festivals/useFestivals.js
@@ -7,6 +7,8 @@ export function useFestivals() {
     const navigate = useNavigate();
 
     // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
 
     // 기존 축제 상세 정보
     const getFestival = useCallback(async (festivalNum) => {
@@ -29,18 +31,15 @@ export function useFestivals() {
             method: 'GET'
 
         }).then((response) => {
-            if (!response.ok) {
-                throw new Error(response.status);
-            }
+            if (!response.ok) { throw new Error(response.status); }
 
             return response.json();
 
-        }).catch((e) => {
-            console.log(e);
-
-        })
+        }).catch((e) => { console.log(e); return undefined; })
     }, [])
 
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
     // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
 
     // 첨푸파일 인코딩
@@ -61,6 +60,8 @@ export function useFestivals() {
         }).catch((e) => { console.log(e); })
     }, [])
 
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
     // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
 
     // 축제 추가/수정
@@ -104,19 +105,37 @@ export function useFestivals() {
     };
 
     // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
+    // ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦ ▦▦▦▦▦▦▦▦▦▦
 
-    // 이미지 첨부 취소
-    const deleteFileFestival = useCallback(async (file) => {
+    // 축제 삭제
+    const deleteFestival = useCallback(async (festivalNum) => {
 
-        return fetch(SERVER_URL + `deleteFileFeatival`, {
-            method: 'DELETE',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(file)
+        const jwt = sessionStorage.getItem('jwt');
 
-        }).then((response) => {
-            if (!response.ok) { throw new Error(response.status); }
+        if (!window.confirm('정말로 삭제하시겠습니까? 다시 복수할수 없습니다!')) { return; }
+        if (jwt === undefined || jwt === '') { alert('로그인이 필요합니다'); return; }
 
-        }).catch((e) => { console.log(e); })
+        // ----- 저장된 축제 이미지 정보와 파일 제거 -----
+        fetch(SERVER_URL + `deleteAllFileFeatival?festivalNum=${festivalNum}`, {
+            method: 'DELETE'
+
+        }).then((res) => {
+            if (!res.json()) { throw new Error(res.status); }
+
+        }).catch((e) => { console.log(e); alert("삭제에 실패하였습니다."); })
+
+        // ----- 저장된 축제 정보 제거 -----
+        fetch(SERVER_URL + `deleteFeatival?festivalNum=${festivalNum}`, {
+            method: 'DELETE'
+
+        }).then((res) => {
+            if (!res.ok) { throw new Error(res.status); }
+
+            alert("삭제가 완료되었습니다.");
+            navigate(`/festivalList`);
+
+        }).catch((e) => { console.log(e); alert("삭제에 실패하였습니다."); })
 
     }, [])
 
@@ -129,6 +148,6 @@ export function useFestivals() {
         submitFestival,
         submitFileFestival,
 
-        deleteFileFestival
+        deleteFestival
     };
 }
